# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API v2",
        "contact": {
            "email": "ybier@esme-solutions.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/data": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "Get list of data",
                "operationId": "5f15d49d6e7531b578bfffa82b89127c",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Data"
                ],
                "summary": "Create a data",
                "operationId": "3db50697065224b1323cfc827d59eb58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ben Data"
                                    },
                                    "point_id": {
                                        "description": "Id of a point",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "description": "Status of the data. 1 or 0",
                                        "type": "string",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/data/{id}": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "Show details of a data",
                "operationId": "b97495cea3fc2607232010fd41669892",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the data to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Data"
                ],
                "summary": "Update a data",
                "operationId": "7f47d6654defdf740d50927c094342b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the data to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Ben Data"
                                            },
                                            "point_id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "status": {
                                                "description": "Status of the data. 1 or 0",
                                                "type": "string",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Data"
                ],
                "summary": "Delete a data",
                "operationId": "e2675f6067c9a7c79110ce192284696e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the data to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deleted"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "clientCredentials": {
                        "authorizationUrl": "http://powerbat-laravel/oauth/authorize",
                        "tokenUrl": "http://powerbat-laravel/oauth/token",
                        "refreshUrl": "http://powerbat-laravel/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}